Обновление 4
Дописаны фильтры без учёта поиска

+++ src/main/java/com/example/testTask/models/VacuumCleaner.java
@@ -20,7 +20,7 @@ public class VacuumCleaner {
 
     @NotNull
     @Column(name = "volume")
-    private double volume;
+    private int volume;
 
     @NotNull
     @Column(name = "number_of_modes")
